Extract redux store configuration from index and allow multiple configurations based on NODE_ENV value
