+++ src/redux/store/configureStore.dev.js
@@ -1,45 +1,32 @@
-import { createStore } from 'redux';
-import rootReducer from '../reducers';
+// @flow
 
-const store = createStore(rootReducer);
+// #region imports
+import { createStore, applyMiddleware } from 'redux';
+import { createLogger } from 'redux-logger';
+import { routerMiddleware } from 'react-router-redux';
+// #region import createHistory from hashHistory or BrowserHistory:
+import createHistory from 'history/createHashHistory';
+// import createHistory            from 'history/createBrowserHistory';
+// #endregion
+import thunkMiddleware from 'redux-thunk';
+import rootReducer from '../reducers/RootReducer';
+import { composeWithDevTools } from 'redux-devtools-extension';
+// #endregion
 
-export default store;
+const loggerMiddleware = createLogger(
+  {
+    level: 'info',
+    collapsed: true,
+  });
 
-// // @flow
-//
-// // #region imports
-// import { createStore, applyMiddleware } from 'redux';
-// import { createLogger } from 'redux-logger';
-// import { routerMiddleware } from 'react-router-redux';
-// // #region import createHistory from hashHistory or BrowserHistory:
-// import createHistory from 'history/createHashHistory';
-// // import createHistory            from 'history/createBrowserHistory';
-// // #endregion
-// import thunkMiddleware from 'redux-thunk';
-// import reducer from '../reducers';
-// import { composeWithDevTools } from 'redux-devtools-extension';
-// // #endregion
-//
-// const loggerMiddleware = createLogger({
-//   level: 'info',
-//   collapsed: true,
-// });
-//
-// export const history = createHistory();
-//
-// // createStore : enhancer
-// // NOTE: use now https://github.com/zalmoxisus/redux-devtools-extension#redux-devtools-extension
-// const enhancer = composeWithDevTools(
-//   applyMiddleware(thunkMiddleware, routerMiddleware(history), loggerMiddleware),
-// );
-//
-// // export default = "redux store"
-// export default function configureStore(initialState) {
-//   const store = createStore(reducer, initialState, enhancer);
-//   if (module.hot) {
-//     module.hot.accept('../reducers', () =>
-//       store.replaceReducer(require('../reducers').default),
-//     );
-//   }
-//   return store;
-// }
+const history = createHistory();
+
+// createStore : enhancer
+// NOTE: use now https://github.com/zalmoxisus/redux-devtools-extension#redux-devtools-extension
+const enhancer = composeWithDevTools(
+  applyMiddleware(thunkMiddleware, routerMiddleware(history), loggerMiddleware),
+);
+
+export default function configureStore(initialState) {
+  return createStore(rootReducer, initialState, enhancer);
+}
